refactor: change properties in keyboardKeys.js; devide app functions in app.js and keyboard.js
