+++ src/index.js
@@ -1,4 +1,4 @@
 import App from './js/app';
 
 const newApp = new App();
-newApp.renderKeys();
+newApp.renderElements();
