+++ src/js/key.js
@@ -3,22 +3,20 @@ import createElementWithProperties from './utils';
 class Key {
   constructor(data, name) {
     this.name = name;
-    this.enValue = data.en.usual;
-    this.enShifted = data.en.shifted;
-    this.ruValue = data.ru.usual;
-    this.ruShifted = data.ru.shifted;
-    this.value = this.enValue;
+    this.language = data.en;
+    this.secondLang = data.ru;
+    this.usualValue = this.language.usual;
+    this.shiftedValue = this.language.shifted;
+    this.value = this.usualValue;
+    this.size = data.size;
     this.systemKey = data.system;
   }
 
   renderKey() {
     const li = createElementWithProperties('li', 'keyboard__item');
     this.button = createElementWithProperties('button', 'keyboard__button');
-    if (this.systemKey) {
-      for (let i = 0; i < this.systemKey.length; i += 1) {
-        this.button.classList.add(this.systemKey[i]);
-      }
-    }
+    if (this.size) this.button.classList.add(this.size);
+    if (this.system) this.button.classList.add(this.system);
     const topSpan = createElementWithProperties('span', 'keyboard__top');
     const bottomSpan = createElementWithProperties('span', 'keyboard__bottom');
     this.buttonLabel = createElementWithProperties('div', 'keyboard__label');
@@ -29,8 +27,8 @@ class Key {
   }
 
   toggleShift(shift) {
-    if (shift) this.value = this.enShifted;
-    if (!shift) this.value = this.enValue;
+    if (shift) this.value = this.shiftedValue;
+    if (!shift) this.value = this.usualValue;
     this.buttonLabel.innerText = this.value;
   }
 }
